[22.05] perlPackages.LWP: keep HTTPDaemon in propagatedBuildInputs
LWP was updated and changed to work around a cross compilation
regression caused by a HTTPDaemon security update, this caused another
regression where derivations expecting LWP to propagate HTTPDaemon
failed.

This commit keeps HTTPDaemon in propagatedBuildInputs for host builds as
some derivations like SOAPLite, Tirex and possibly more expect LWP to
propagate it. But not for cross builds since they are not able to build
HTTPDaemon.